Fix missing link issues in site generation
The `metanorma` requires us to have `rxl` as an extension for some
standards, and if missing then it can't generate the collection in
proper way. This was one the issue for missing link.

This commit should fix the issue and it should link the internal
documents properly.

One thing to note, there is still one issue with how it handles
the `RFC`, but that is something we will address soon.

Related: https://github.com/metanorma/metanorma-cli/issues/297